Switch to github.com/klauspost/compress/pgzip
From casual experiments, I found that there was a noticeable drag with
the gzipping part of the code. With one Go binary, I had rough timings
as follows for that section:

- github.com/klauspost/pgzip: ~80ms
- github.com/klauspost/compress/gzip: ~450ms
- compress/gzip: ~1250ms

There appears to be no penalty to using pgzip in this context (as tested
by setting GOMAXPROCS=1).

This commit switches from compress/gzip to github.com/klauspost/pgzip.